+++ rust/main/utils/run-locally/src/invariants/termination_invariants.rs
@@ -1,3 +1,4 @@
+use std::collections::HashMap;
 use std::fs::File;
 
 use crate::config::Config;
@@ -212,6 +213,14 @@ pub fn relayer_termination_invariants_met(
         total_messages_expected + non_matching_igp_message_count + double_insertion_message_count,
     );
 
+    let dropped_tasks = fetch_metric(
+        RELAYER_METRICS_PORT,
+        "hyperlane_tokio_dropped_tasks",
+        &hashmap! {"agent" => "relayer"},
+    )?;
+
+    assert_eq!(dropped_tasks.first().unwrap(), 0);
+
     if !relayer_balance_check(starting_relayer_balance)? {
         return Ok(false);
     }
@@ -291,5 +300,19 @@ pub fn relayer_balance_check(starting_relayer_balance: f64) -> eyre::Result<bool
         );
         return Ok(false);
     }
+    Ok(true)
+}
+
+#[allow(dead_code)]
+pub fn provider_metrics_invariant_met(
+    relayer_port: &str,
+    expected_request_count: u32,
+    filter_hashmap: &HashMap<&str, &str>,
+) -> eyre::Result<bool> {
+    let request_count = fetch_metric(relayer_port, "hyperlane_request_count", filter_hashmap)?
+        .iter()
+        .sum::<u32>();
+
+    assert!(request_count > expected_request_count);
     Ok(true)
 }