+++ typescript/infra/src/config/gas-oracle.ts
@@ -171,16 +171,23 @@ function getMinUsdCost(local: ChainName, remote: ChainName): number {
     // aren't accounted for directly in the gas price.
     arbitrum: 0.5,
     ancient8: 0.5,
-    base: 0.25,
     blast: 0.5,
     bob: 0.5,
-    fraxtal: 0.5,
     linea: 0.5,
     mantapacific: 0.5,
     mantle: 0.5,
-    mode: 0.5,
-    optimism: 0.25,
     polygonzkevm: 0.5,
+
+    // op stack chains
+    base: 0.2,
+    fraxtal: 0.2,
+    lisk: 0.2,
+    mode: 0.2,
+    optimism: 0.2,
+    soneium: 0.2,
+    superseed: 0.2,
+    unichain: 0.2,
+
     // Scroll is more expensive than the rest due to higher L1 fees
     scroll: 1.5,
     taiko: 0.5,